==> 01-Plan_your_trip_with_kayak.ipynb <==
# # Projet : Planifier mes prochaines vacances 

### Utilisons les API et le web-scraping pour déterminer quels seraient les plus beaux endroits à visiter pour des vacances de dernière minute la semaine prochaine !

## Selon <a href="https://one-week-in.com/35-cities-to-visit-in-france/" target="_blank">ce site web</a>, voici les 35 meilleurs endroits à visiter en France en 2020 :

 ```python
["Mont Saint Michel",
"St Malo",
"Bayeux",
"Le Havre",
"Rouen",
"Paris",
"Amiens",
"Lille",
"Strasbourg",
"Chateau du Haut Koenigsbourg",
"Colmar",
"Eguisheim",
"Besancon",
"Dijon",
"Annecy",
"Grenoble",
"Lyon",
"Gorges du Verdon",
"Bormes les Mimosas",
"Cassis",
"Marseille",
"Aix en Provence",
"Avignon",
"Uzes",
"Nimes",
"Aigues Mortes",
"Saintes Maries de la mer",
"Collioure",
"Carcassonne",
"Ariege",
"Toulouse",
"Montauban",
"Biarritz",
"Bayonne",
"La Rochelle"]
```

### Utilisez <a href="https://nominatim.org/" target="_blank">https://nominatim.org/</a> pour obtenir les coordonnées gps de toutes les villes (aucun abonnement requis).

### Documentation : <a href="https://nominatim.org/release-docs/develop/api/Search/" target="_blank">https://nominatim.org/release-docs/develop/api/Search/</a>

### 3. Déterminez la liste des villes où le temps sera le plus agréable dans les 7 prochains jours.

Par exemple, vous pouvez utiliser les valeurs de `daily.pop` et `daily.rain` pour calculer le volume de pluie attendu dans les 7 prochains jours... Mais ce n'est qu'un exemple, en fait vous pouvez avoir des opinions différentes sur ce que serait un "beau temps" 😎. Peut-être que le critère le plus important pour vous est la température ou l'humidité, alors n'hésitez pas à changer les règles !

### 4. Enregistrez tous les résultats dans un fichier .csv, vous l'utiliserez plus tard 😉 .

Vous pouvez enregistrer toutes les informations qui vous semblent importantes ! N'oubliez pas de sauvegarder le nom des villes, et aussi de créer une colonne contenant un identifiant unique (id) de chaque ville (c'est important pour la suite du projet).

### 5. Utilisez plotly pour afficher les meilleures destinations sur une carte.

# Import des librairies

import json
import requests
import pandas as pd
import plotly.express as px
from plotly import graph_objs as go

!pip install plotly
!pip install boto3
!pip install python-box
!pip install pyyaml
#!jupyter labextension install jupyterlab-plotly@4.8.1

city_names = [
    "Mont Saint Michel",
    "St Malo",
    "Bayeux",
    "Le Havre",
    "Rouen",
    "Paris",
    "Amiens",
    "Lille",
    "Strasbourg",
    "Chateau du Haut Koenigsbourg",
    "Colmar",
    "Eguisheim",
    "Besancon",
    "Dijon",
    "Annecy",
    "Grenoble",
    "Lyon",
    "Gorges du Verdon",
    "Bormes les Mimosas",
    "Cassis",
    "Marseille",
    "Aix en Provence",
    "Avignon",
    "Uzes",
    "Nimes",
    "Aigues Mortes",
    "Saintes Maries de la mer",
    "Collioure",
    "Carcassonne",
    "Foix",
    "Toulouse",
    "Montauban",
    "Biarritz",
    "Bayonne",
    "La Rochelle"
]

dataset = pd.DataFrame(columns=['city_id' , 'name', 'latitude', 'longitude', 'main_weather', 'expected_rain', 'day_temperature'])

#  IDENTIFIANT D'APPLICATION (VOUS DEVEZ VOUS ABONNER À OPENWEATHERMAP)

API_KEY = "755cc401942e8750f3ee4a975b76fd81"

for i in range(len(city_names)):
    
    print("Making requests for {}".format(city_names[i]))
    
  
    # Utiliser l'api nominatim pour obtenir les coordonnées GPS de la ville
    
    NOMINATIM_URL = 'https://nominatim.openstreetmap.org/search'
    params = {
        "city": city_names[i],
        "country": "France",
        "format": "json"
    }
    
    r = requests.get(NOMINATIM_URL, params=params)
    res = r.json()
    
    # ex: Gorges du Verdon
    if len(res) < 1:
        params = {
            "q": city_names[i],
            "country": "France",
            "format": "json"
        }

        r = requests.get(NOMINATIM_URL, params=params)
        res = r.json()    
    
    dataset.loc[i,'city_id'] = i
    dataset.loc[i,'name'] = city_names[i]
    dataset.loc[i, 'latitude'] = res[0]['lat']
    dataset.loc[i, 'longitude'] = res[0]['lon']
    
  
    # Utiliser l'api openweathermap pour obtenir le temps pour les 7 prochains jours
    
    params = {
        "lat": dataset.loc[i, 'latitude'],
        "lon": dataset.loc[i, 'longitude'],
        "exclude": "current,minutely,hourly",
        "units": "metric",
        "appid": API_KEY,
    }

    r = requests.get('https://api.openweathermap.org/data/2.5/onecall', params=params)
    res = r.json()
    
   
    # Calculer le volume attendu de pluie
    
    expected_rain = 0
    # print(res)
    for d in res['daily']:
        if 'rain' in d.keys():
            expected_rain += d['pop'] * d['rain']
            
 
    # Calculer la température moyenne du jour
    
    temperatures = pd.Series([d['temp']['day'] for d in res['daily']])
    mean_temperature = temperatures.mean()


    # Extrait du temps le plus probable
    
    weathers = pd.Series([d['weather'][0]['main'] for d in res['daily']])
    main_weather = weathers.mode()[0]
    
    dataset.loc[i,'main_weather'] = main_weather
    dataset.loc[i,'expected_rain'] = expected_rain
    dataset.loc[i,'day_temperature'] = mean_temperature


 len(city_names)

# Mettre les villes récupérées dans un cadre de données

dataset = dataset.astype({'city_id': int,
                          'latitude': float,
                          'longitude': float,
                          'expected_rain': float,
                          'day_temperature': float,
                         })

dataset.loc[:, 'rank'] = dataset['expected_rain'].rank(method='min').astype(int)
dataset.loc[:, 'inverted_rank'] = dataset['expected_rain'].rank(method='min', ascending=False).astype(int)

dataset = dataset.sort_values(by=['expected_rain', 'day_temperature'], ascending=[True, False]).reset_index(drop=True)
display(dataset)

# Les meilleurs endroits pour un voyage la semaine prochaine

print('Les meilleurs endroits pour un voyage la semaine prochaine sont : ')
print()

for i, row in dataset.loc[dataset['rank']==1,:].iterrows():
    print("{} -- Mostly {} with temperature {} °C".format(row['name'], row['main_weather'], row['day_temperature']))

# Enregistrement du fichier 

dataset.to_csv('res/1_destinations.csv', index=False)

# Visualisation sur une carte destinations

fig = px.scatter_mapbox(dataset, lat="latitude", lon="longitude", hover_name = 'name', zoom = 4,
                        hover_data = ['main_weather', 'expected_rain', 'day_temperature'], 
                        color = 'day_temperature', color_continuous_scale = 'Bluered', size = 'inverted_rank',
                        mapbox_style="carto-positron")
fig.show('iframe')

dataset.to_csv('res/1_destinations.csv', index=False)
