==> 02-Plan_your_trip_with_kayak.ipynb <==
# Projet : planifier mes prochaines vacances 

### Créons un script qui permet d'obtenir quelques informations sur tous les hôtels d'une ville donnée sur www.booking.com 🧙.

### on va utiliser Scrapy pour scrapper ces informations:

### * Le nom de l'hôtel,
### * L'url de sa page booking.com,
### * Ses coordonnées : latitude et longitude,
### * La note donnée par les utilisateurs du site,
### * La description textuelle de l'hôtel.

### Ensuite, on va exécuter ce script pour plusieurs villes de la liste. Assurez-vous de sauvegarder les résultats dans des fichiers différents pour chaque ville et que le nom de la ville est stocké dans le nom du fichier (pour des raisons ultérieures 😉 ).

!pip install Scrapy

import os
import logging

import scrapy
from scrapy.crawler import CrawlerProcess

destination_name = "Aix en Provence"  

class Hotels(scrapy.Spider):
    # from plotly import graph_objs as go
    name = "hotels"

    # Starting URL
    start_urls = ['https://www.booking.com/index.fr.html']

    # Parse function for login        # Fonction d'analyse pour le login
    
    def parse(self, response):
        # FormRequest used to login   # FormRequest utilisé pour se connecter
        return scrapy.FormRequest.from_response(
            response,
            formdata={'ss': destination_name},
            callback=self.after_search
        )

    # Callback used after login        # Callback utilisé après la connexion
    def after_search(self, response):
        
        hotels = response.css('.sr_item')

        for h in hotels:
            yield {
                'name': h.css('.sr-hotel__name::text').get(),
                'url': "https://www.booking.com" + h.css('.hotel_name_link').attrib["href"],
                'coords': h.css('.sr_card_address_line a').attrib["data-coords"],
                'score': h.css('.bui-review-score__badge::text').get(),
                'description': h.css('.hotel_desc::text').get()
                
            }
        
        
        # Select the NEXT button and store it in next_page    # Sélectionnez le bouton NEXT et enregistrez-le dans next_page
        try:
            next_page = response.css('a.paging-next').attrib["href"]
        except KeyError:
            logging.info('No next page. Terminating crawling process.')
        else:
            yield response.follow(next_page, callback=self.after_search)

filename = "2_hotels_" + destination_name.replace(" ", "-") + ".json"

if filename in os.listdir('res/'):
    os.remove('res/' + filename)

process = CrawlerProcess(settings = {
    'USER_AGENT': 'Chrome/84.0 (compatible; MSIE 7.0; Windows NT 5.1)',
    'LOG_LEVEL': logging.INFO,
    "FEEDS": {
        'res/' + filename: {"format": "json"},
    }
})

process.crawl(Hotels)
process.start()
