==> 04-Plan_your_trip_with_kayak.ipynb <==
# Projet : planifier mes prochaines vacances

### Mettons notre ensemble de données dans une base de données postgreSQL sur Amazon RDS ! Ensuite, vous pourrez utiliser la syntaxe SQL pour faire des requêtes et décider où vous irez en vacances.

!pip install psycopg2-binary

from box import Box
import yaml # Seulement nécessaire pour les différents chargeurs
import pandas as pd

!pwd

conf = Box.from_yaml(filename="./config.yaml")

## 1 Lisez les fichiers `.csv` contenant des informations sur les villes et les hôtels :

import pandas as pd
from sqlalchemy import create_engine

cities = pd.read_csv('https://m03-jedha-project-040121.s3.amazonaws.com/cities.csv')  # pd.read_csv('res/1_destinations.csv')
hotels = pd.read_csv('https://m03-jedha-project-040121.s3.amazonaws.com/cities.csv')  # pd.read_csv('res/3_hotels.csv')

cities.head()

hotels.head()

## 2. Créez et configurez une instance RDS dans notre compte AWS.
## 3. Utilisez SQLAlchemy pour créer un moteur qui est connecté à la base de données distante :

# Replace HOSTNAME by AWS RDS Endpoint, USERNAME and PASSWORD by its credentials
# Remplacer HOSTNAME par AWS RDS Endpoint, USERNAME et PASSWORD par ses informations d'identification.

engine = create_engine("postgresql+psycopg2://USERNAME:PASSWORD@HOSTNAME/postgres", echo=True)

## 4. Créez deux tables dans votre base de données distante : `villes` et `hôtels`, chacune contenant les informations des différents fichiers `.csv` que vous avez sauvegardés la dernière fois.

cities.to_sql("cities", engine)
hotels.to_sql("hotels", engine)

## 5. Facultatif : utilisez PGAdmin pour explorer la base de données.

## 6. Utilisez la syntaxe SQL pour effectuer une requête qui vous permettra d'obtenir la liste des 30 meilleurs endroits à visiter (parmi toutes les villes et tous les hôtels). 

statement = "SELECT city_name, main_weather, day_temperature, hotels.name, description, score, url \
             FROM cities \
             INNER JOIN hotels \
             ON cities.city_id = hotels.city_id \
             WHERE rank=1 AND score IS NOT NULL \
             ORDER BY score DESC \
             LIMIT 30"

best_places = pd.read_sql(statement, engine)

display(best_places)
